Test-slim Phase 3, cluster 1: module supply chain
13 phase-named files + the shard boundary file merge into 14
module-named targets (registry, validator, modules API, install API,
publishers, signing, sandbox, mcp_host, trust ceremony, grants
lifecycle, dependency resolver, grants API, grant scope schema, matter
context). Duplicate ceremony scaffolds share fixtures; same-behaviour
assertions deduped one-per-failure-mode; governance files renamed only.

The CI shard split is now computed (sorted glob, cut at half the
cumulative line count; shard-a/shard-b) — the hard-coded boundary
filename this cluster deletes is no longer load-bearing.

Suite: 810 passed, known 3-failure macOS sandbox baseline, zero
collection errors.
