+++ backend/tests/test_dependency_resolver.py
