+++ backend/tests/test_grants_lifecycle.py
@@ -1,13 +1,13 @@
-"""Round-2 Reviewer fix tests (Phase 3 + Phase 4).
+"""Grants lifecycle — permission expansion detection + round-2 regression
+fixes.
 
-Four findings from the combined Phase 3+4 ratification review:
+Merged from test_phase4_grants_lifecycle.py +
+test_phase3_phase4_round2_fixes.py. Round-2 findings covered:
 
 P1#1 — trust ceremony grant could skip straight to enabled
 P1#2 — update diff missed advice-tier expansion in raw manifests
 P1#3 — dependency resolver never called from install/update
 P2#4 — duplicate persist on retried final grant
-
-Each finding gets a focused regression test here.
 """
 
 from __future__ import annotations
@@ -17,7 +17,11 @@ import uuid
 import pytest
 from sqlalchemy import select
 
-from app.core.grants_lifecycle import detect_expansion, requires_reprompt
+from app.core.grants_lifecycle import (
+    ExpansionReport,
+    detect_expansion,
+    requires_reprompt,
+)
 from app.core.trust_ceremony import (
     CeremonyState,
     InvalidCeremonyTransition,
@@ -26,6 +30,46 @@ from app.core.trust_ceremony import (
     start_ceremony,
 )
 from app.models import InstalledModule, User
+
+
+# ---------------------------------------------------------------------------
+# Shared scaffolding
+# ---------------------------------------------------------------------------
+
+
+def _snapshot(
+    *,
+    reads=None,
+    writes=None,
+    advice_tier_max="draft_advice",
+    external_network=False,
+    destinations=None,
+    gates=None,
+    model_access="none",
+) -> dict:
+    return {
+        "advice_tier_max": advice_tier_max,
+        "data_movement": {
+            "external_destinations": destinations or [],
+        },
+        "gates": gates or [],
+        "capabilities": [
+            {
+                "id": "default",
+                "kind": "skill",
+                "scope": "matter",
+                "reads": reads or [],
+                "writes": writes or [],
+                "model_access": model_access,
+                "external_network": external_network,
+                "data_movement": {
+                    "external_destinations": destinations or [],
+                },
+                "gates": gates or [],
+                "advice_tier_max": advice_tier_max,
+            }
+        ],
+    }
 
 
 def _verified_manifest(module_id="legalise.r2-test", version="1.0.0", **overrides) -> dict:
@@ -79,8 +123,184 @@ async def _make_user(db_session, *, is_superuser: bool = False) -> User:
     return user
 
 
+async def _login_admin(client) -> str:
+    """Register a user, flip to superuser, log in. Returns the email."""
+    email = f"r2-admin-{uuid.uuid4().hex[:8]}@example.com"
+    password = "round2-2026"
+    await client.post(
+        "/auth/register", json={"email": email, "password": password}
+    )
+
+    from app.main import app
+
+    factory = app.state.session_factory
+    async with factory() as session:
+        u = await session.scalar(select(User).where(User.email == email))
+        u.is_superuser = True
+        await session.commit()
+
+    await client.post(
+        "/auth/login",
+        data={"username": email, "password": password},
+        headers={"Content-Type": "application/x-www-form-urlencoded"},
+    )
+    return email
+
+
+# ---------------------------------------------------------------------------
+# detect_expansion / requires_reprompt (pure unit)
+# ---------------------------------------------------------------------------
+
+
+def test_identical_snapshots_no_expansion() -> None:
+    a = _snapshot(reads=["matter.read"], writes=["citation.write"])
+    b = _snapshot(reads=["matter.read"], writes=["citation.write"])
+    report = detect_expansion(a, b)
+    assert report.any_expansion is False
+    assert requires_reprompt(report) is False
+
+
+def test_new_read_capability_is_expansion() -> None:
+    old = _snapshot(reads=["matter.read"])
+    new = _snapshot(reads=["matter.read", "document.body.read"])
+    report = detect_expansion(old, new)
+    assert "document.body.read" in report.reads_added
+    assert requires_reprompt(report)
+
+
+def test_new_write_capability_is_expansion() -> None:
+    old = _snapshot(writes=[])
+    new = _snapshot(writes=["citation.write"])
+    report = detect_expansion(old, new)
+    assert "citation.write" in report.writes_added
+    assert requires_reprompt(report)
+
+
+def test_tier_raise_is_expansion() -> None:
+    old = _snapshot(advice_tier_max="draft_advice")
+    new = _snapshot(advice_tier_max="supervised_legal_advice")
+    report = detect_expansion(old, new)
+    assert report.tier_raised == ("draft_advice", "supervised_legal_advice")
+    assert requires_reprompt(report)
+
+
+def test_tier_lower_is_not_expansion() -> None:
+    old = _snapshot(advice_tier_max="supervised_legal_advice")
+    new = _snapshot(advice_tier_max="draft_advice")
+    report = detect_expansion(old, new)
+    assert report.tier_raised is None
+    assert not requires_reprompt(report)
+
+
+def test_external_network_flip_is_expansion() -> None:
+    old = _snapshot(external_network=False)
+    new = _snapshot(external_network=True, destinations=["api.example.com"])
+    report = detect_expansion(old, new)
+    assert report.external_network_added is True
+    assert "api.example.com" in report.new_destinations
+    assert requires_reprompt(report)
+
+
+def test_new_external_destination_is_expansion() -> None:
+    old = _snapshot(external_network=True, destinations=["api.a.com"])
+    new = _snapshot(
+        external_network=True, destinations=["api.a.com", "api.b.com"]
+    )
+    report = detect_expansion(old, new)
+    assert report.external_network_added is False  # already on
+    assert "api.b.com" in report.new_destinations
+    assert "api.a.com" not in report.new_destinations
+    assert requires_reprompt(report)
+
+
+def test_new_gate_is_expansion() -> None:
+    old = _snapshot(gates=["privilege_posture"])
+    new = _snapshot(gates=["privilege_posture", "advice_boundary"])
+    report = detect_expansion(old, new)
+    assert "advice_boundary" in report.new_gates_added
+    assert requires_reprompt(report)
+
+
+def test_gate_removed_alone_not_expansion() -> None:
+    """Gate removal is recorded but doesn't trigger re-prompt by
+    Phase 4 policy (only additions expand permissions)."""
+    old = _snapshot(gates=["privilege_posture", "advice_boundary"])
+    new = _snapshot(gates=["privilege_posture"])
+    report = detect_expansion(old, new)
+    assert "advice_boundary" in report.new_gates_removed
+    # Phase 4 policy: any_expansion picks up added but not removed
+    # alone. requires_reprompt returns False here.
+    assert report.any_expansion is False
+
+
+def test_model_access_raise_is_expansion() -> None:
+    old = _snapshot(model_access="none")
+    new = _snapshot(model_access="required")
+    report = detect_expansion(old, new)
+    assert report.model_access_raised == ("none", "required")
+    assert requires_reprompt(report)
+
+
+def test_to_dict_serialisable() -> None:
+    """ExpansionReport.to_dict produces JSON-serialisable output for
+    the API response."""
+    import json
+
+    old = _snapshot()
+    new = _snapshot(advice_tier_max="supervised_legal_advice")
+    report = detect_expansion(old, new)
+    d = report.to_dict()
+    json.dumps(d)  # must not raise
+    assert d["tier_raised"] == {
+        "from": "draft_advice",
+        "to": "supervised_legal_advice",
+    }
+
+
+# ---------------------------------------------------------------------------
+# P1#2 — advice-tier expansion in raw manifests is now detected
+# ---------------------------------------------------------------------------
+
+
+def test_detect_expansion_picks_up_tier_increase_in_capabilities() -> None:
+    """Round-2 P1#2: when the new snapshot is a raw v2 manifest (no
+    top-level advice_tier_max key, only per-capability), the diff
+    must still pick up the tier increase."""
+    # Old snapshot is aggregated (built by build_permission_card).
+    old = {
+        "advice_tier_max": "draft_advice",
+        "capabilities": [{"advice_tier_max": "draft_advice"}],
+    }
+    # New snapshot is the raw manifest shape — no top-level rollup.
+    new = {
+        "capabilities": [{"advice_tier_max": "supervised_legal_advice"}],
+    }
+    report = detect_expansion(old, new)
+    assert report.tier_raised == ("draft_advice", "supervised_legal_advice")
+    assert requires_reprompt(report)
+
+
+def test_detect_expansion_handles_multi_capability_tier() -> None:
+    """If any capability in the new manifest declares a higher tier
+    than the old aggregate, it counts as expansion."""
+    old = {
+        "advice_tier_max": "factual_extraction",
+        "capabilities": [{"advice_tier_max": "factual_extraction"}],
+    }
+    new = {
+        "capabilities": [
+            {"advice_tier_max": "factual_extraction"},
+            {"advice_tier_max": "draft_advice"},
+        ],
+    }
+    report = detect_expansion(old, new)
+    assert report.tier_raised == ("factual_extraction", "draft_advice")
+
+
 # ---------------------------------------------------------------------------
 # P1#1 — grant can no longer skip the ceremony
+# (grant-from-GRANTED happy path is pinned by
+#  test_trust_ceremony.test_verified_path_reaches_enabled_in_three_advances)
 # ---------------------------------------------------------------------------
 
 
@@ -138,74 +358,6 @@ async def test_grant_from_publisher_checked_raises(db_session) -> None:
         )
 
 
-@pytest.mark.asyncio
-async def test_grant_from_granted_succeeds(db_session) -> None:
-    """Round-2 P1#1: grant IS allowed from GRANTED."""
-    clear_ceremonies()
-    user = await _make_user(db_session)
-    ceremony = await start_ceremony(
-        db_session,
-        manifest=_verified_manifest(module_id="legalise.r2-grant-ok"),
-        actor_user_id=user.id,
-    )
-    # Fast path: 3 trusts to reach GRANTED.
-    for _ in range(3):
-        ceremony = await advance_ceremony(
-            db_session,
-            ceremony_id=ceremony.id,
-            action="trust",
-            actor_user_id=user.id,
-        )
-    assert ceremony.state is CeremonyState.GRANTED
-    enabled = await advance_ceremony(
-        db_session,
-        ceremony_id=ceremony.id,
-        action="grant",
-        actor_user_id=user.id,
-    )
-    assert enabled.state is CeremonyState.ENABLED
-
-
-# ---------------------------------------------------------------------------
-# P1#2 — advice-tier expansion in raw manifests is now detected
-# ---------------------------------------------------------------------------
-
-
-def test_detect_expansion_picks_up_tier_increase_in_capabilities() -> None:
-    """Round-2 P1#2: when the new snapshot is a raw v2 manifest (no
-    top-level advice_tier_max key, only per-capability), the diff
-    must still pick up the tier increase."""
-    # Old snapshot is aggregated (built by build_permission_card).
-    old = {
-        "advice_tier_max": "draft_advice",
-        "capabilities": [{"advice_tier_max": "draft_advice"}],
-    }
-    # New snapshot is the raw manifest shape — no top-level rollup.
-    new = {
-        "capabilities": [{"advice_tier_max": "supervised_legal_advice"}],
-    }
-    report = detect_expansion(old, new)
-    assert report.tier_raised == ("draft_advice", "supervised_legal_advice")
-    assert requires_reprompt(report)
-
-
-def test_detect_expansion_handles_multi_capability_tier() -> None:
-    """If any capability in the new manifest declares a higher tier
-    than the old aggregate, it counts as expansion."""
-    old = {
-        "advice_tier_max": "factual_extraction",
-        "capabilities": [{"advice_tier_max": "factual_extraction"}],
-    }
-    new = {
-        "capabilities": [
-            {"advice_tier_max": "factual_extraction"},
-            {"advice_tier_max": "draft_advice"},
-        ],
-    }
-    report = detect_expansion(old, new)
-    assert report.tier_raised == ("factual_extraction", "draft_advice")
-
-
 # ---------------------------------------------------------------------------
 # P1#3 — install rejects manifests with missing dependencies
 # ---------------------------------------------------------------------------
@@ -215,26 +367,7 @@ def test_detect_expansion_handles_multi_capability_tier() -> None:
 async def test_install_rejects_missing_dependency(client) -> None:
     """Round-2 P1#3: start_install_endpoint must run
     resolve_dependencies and reject when unsatisfied."""
-    # Register an admin.
-    email = f"r2-dep-{uuid.uuid4().hex[:8]}@example.com"
-    password = "round2-2026"
-    await client.post(
-        "/auth/register", json={"email": email, "password": password}
-    )
-
-    # Flip to superuser via a fresh session.
-    from app.main import app
-    factory = app.state.session_factory
-    async with factory() as session:
-        u = await session.scalar(select(User).where(User.email == email))
-        u.is_superuser = True
-        await session.commit()
-
-    await client.post(
-        "/auth/login",
-        data={"username": email, "password": password},
-        headers={"Content-Type": "application/x-www-form-urlencoded"},
-    )
+    await _login_admin(client)
 
     # Manifest declares a dependency that doesn't exist.
     manifest = _verified_manifest(module_id="legalise.needs-dep")
@@ -257,27 +390,8 @@ async def test_install_rejects_missing_dependency(client) -> None:
 @pytest.mark.asyncio
 async def test_update_rejects_missing_dependency(client) -> None:
     """Round-2 P1#3: update_module_endpoint also runs the resolver."""
-    from app.core.trust_ceremony import clear_ceremonies
-
     clear_ceremonies()
-    email = f"r2-updep-{uuid.uuid4().hex[:8]}@example.com"
-    password = "round2-2026"
-    await client.post(
-        "/auth/register", json={"email": email, "password": password}
-    )
-
-    from app.main import app
-    factory = app.state.session_factory
-    async with factory() as session:
-        u = await session.scalar(select(User).where(User.email == email))
-        u.is_superuser = True
-        await session.commit()
-
-    await client.post(
-        "/auth/login",
-        data={"username": email, "password": password},
-        headers={"Content-Type": "application/x-www-form-urlencoded"},
-    )
+    await _login_admin(client)
 
     # Install v1.0.0 (no deps).
     manifest_v1 = _verified_manifest(module_id="legalise.dep-update", version="1.0.0")
@@ -323,27 +437,8 @@ async def test_repeated_grant_does_not_double_insert(client) -> None:
     on the transition INTO enabled. A retry/poll of the final grant
     must not attempt a second insert (which would 500 on the
     UNIQUE (module_id, version) constraint)."""
-    from app.core.trust_ceremony import clear_ceremonies
-
     clear_ceremonies()
-    email = f"r2-idemp-{uuid.uuid4().hex[:8]}@example.com"
-    password = "round2-2026"
-    await client.post(
-        "/auth/register", json={"email": email, "password": password}
-    )
-
-    from app.main import app
-    factory = app.state.session_factory
-    async with factory() as session:
-        u = await session.scalar(select(User).where(User.email == email))
-        u.is_superuser = True
-        await session.commit()
-
-    await client.post(
-        "/auth/login",
-        data={"username": email, "password": password},
-        headers={"Content-Type": "application/x-www-form-urlencoded"},
-    )
+    await _login_admin(client)
 
     manifest = _verified_manifest(module_id="legalise.idemp-test", version="1.0.0")
     install_start = await client.post(
@@ -376,6 +471,9 @@ async def test_repeated_grant_does_not_double_insert(client) -> None:
     assert r2.json()["state"] == "enabled"
 
     # Verify exactly one InstalledModule row.
+    from app.main import app
+
+    factory = app.state.session_factory
     async with factory() as session:
         rows = (
             await session.scalars(
@@ -397,26 +495,10 @@ async def test_repeated_grant_does_not_double_insert(client) -> None:
 async def test_unknown_ceremony_action_rejected_at_api(client) -> None:
     """Round-2 residual P2: AdvanceCeremonyRequest.action is now a
     Literal. ``{"action":"banana"}`` must return 422 from FastAPI
-    validation, never advance the ceremony as ``trust``."""
+    validation, never advance the ceremony as ``trust``. (The audit
+    emission for this path is pinned in test_trust_ceremony.)"""
     clear_ceremonies()
-    email = f"r2-action-{uuid.uuid4().hex[:8]}@example.com"
-    password = "round2-2026"
-    await client.post(
-        "/auth/register", json={"email": email, "password": password}
-    )
-
-    from app.main import app
-    factory = app.state.session_factory
-    async with factory() as session:
-        u = await session.scalar(select(User).where(User.email == email))
-        u.is_superuser = True
-        await session.commit()
-
-    await client.post(
-        "/auth/login",
-        data={"username": email, "password": password},
-        headers={"Content-Type": "application/x-www-form-urlencoded"},
-    )
+    await _login_admin(client)
 
     install_start = await client.post(
         "/api/modules/install",
