+++ backend/tests/test_registry.py
