+++ backend/tests/test_registry_validator.py
