+++ backend/tests/test_trust_ceremony.py
@@ -1,4 +1,10 @@
-"""Phase 3 — trust ceremony state machine tests."""
+"""Trust ceremony — state machine, rejection audit, dead-state pins.
+
+Merged from test_phase3_trust_ceremony.py +
+test_phase5_ceremony_rejection_audit.py + test_phase5_carryover_tidy.py
+(the carryover file's app-wide hygiene greps moved to
+test_migration_discipline.py).
+"""
 
 from __future__ import annotations
 
@@ -66,7 +72,7 @@ def _unverified_manifest(**overrides) -> dict:
 async def _make_user(db_session) -> User:
     user = User(
         id=uuid.uuid4(),
-        email=f"p3-tc-{uuid.uuid4().hex[:8]}@example.com",
+        email=f"tc-{uuid.uuid4().hex[:8]}@example.com",
         hashed_password="x" * 32,
         is_active=True,
         is_verified=True,
@@ -78,7 +84,7 @@ async def _make_user(db_session) -> User:
 
 
 # ---------------------------------------------------------------------------
-# Pure unit
+# Pure unit — permission card
 # ---------------------------------------------------------------------------
 
 
@@ -209,25 +215,8 @@ async def test_unverified_path_walks_seven_states(db_session) -> None:
 
 
 @pytest.mark.asyncio
-async def test_reject_terminates_ceremony(db_session) -> None:
-    clear_ceremonies()
-    user = await _make_user(db_session)
-    ceremony = await start_ceremony(
-        db_session,
-        manifest=_verified_manifest(),
-        actor_user_id=user.id,
-    )
-    rejected = await advance_ceremony(
-        db_session,
-        ceremony_id=ceremony.id,
-        action="reject",
-        actor_user_id=user.id,
-    )
-    assert rejected.state is CeremonyState.REJECTED_BY_USER
-
-
-@pytest.mark.asyncio
 async def test_terminal_state_does_not_advance_further(db_session) -> None:
+    """reject terminates the ceremony; further advances are no-ops."""
     clear_ceremonies()
     user = await _make_user(db_session)
     ceremony = await start_ceremony(
@@ -303,3 +292,158 @@ def test_clear_ceremonies_wipes_state() -> None:
     # Cannot directly assert _CEREMONIES is empty since it's private,
     # but get_ceremony for any uuid returns None.
     assert get_ceremony(uuid.uuid4()) is None
+
+
+# ---------------------------------------------------------------------------
+# module.ceremony.rejected audit emission (HTTP-level)
+#
+# Two paths must emit the audit row:
+# 1. InvalidCeremonyTransition raised inside advance_ceremony and
+#    translated to HTTP 409 by api.modules.advance_install_endpoint.
+# 2. Pydantic RequestValidationError raised when the request body
+#    carries an action that isn't in Literal["trust","reject","grant"],
+#    translated to HTTP 422 by the global handler in main.py.
+# ---------------------------------------------------------------------------
+
+
+@pytest.fixture
+def captured_audit_failures(monkeypatch):
+    """Capture every audit_failure call without writing to DB.
+
+    Reason: audit_failure opens an independent committed session for
+    production durability through rollback. In tests, the user only
+    exists inside the outer SAVEPOINT — the independent session
+    cannot see it. Pattern matches
+    test_storage_failure_envelopes.py.
+
+    Returns a list that the test reads to assert on emissions.
+    """
+    from app.core import api as api_module
+
+    captured: list[dict] = []
+
+    async def _capture(session, action, **kwargs):
+        captured.append({"action": action, **kwargs})
+
+    monkeypatch.setattr(api_module, "audit_failure", _capture)
+    return captured
+
+
+async def _login_admin(client) -> tuple[str, str]:
+    email = f"tc-rejaudit-{uuid.uuid4().hex[:8]}@example.com"
+    password = "trust-ceremony-2026"
+    await client.post(
+        "/auth/register", json={"email": email, "password": password}
+    )
+    from app.main import app
+
+    factory = app.state.session_factory
+    async with factory() as session:
+        u = await session.scalar(select(User).where(User.email == email))
+        u.is_superuser = True
+        await session.commit()
+    await client.post(
+        "/auth/login",
+        data={"username": email, "password": password},
+        headers={"Content-Type": "application/x-www-form-urlencoded"},
+    )
+    return email, password
+
+
+@pytest.mark.asyncio
+async def test_invalid_ceremony_transition_emits_audit_row(
+    client, captured_audit_failures
+) -> None:
+    """grant from DISCOVERED → 409 + module.ceremony.rejected row."""
+    clear_ceremonies()
+    email, _ = await _login_admin(client)
+
+    install = await client.post(
+        "/api/modules/install",
+        json={
+            "source": "manifest",
+            "manifest": _verified_manifest(id="legalise.r2-rejaudit"),
+        },
+    )
+    assert install.status_code == 201, install.text
+    ceremony_id = install.json()["ceremony_id"]
+
+    resp = await client.post(
+        f"/api/modules/install/{ceremony_id}/advance",
+        json={"action": "grant"},
+    )
+    assert resp.status_code == 409
+
+    # The endpoint's InvalidCeremonyTransition branch routes through
+    # the captured audit_failure helper. Exactly one row should land
+    # with the canonical shape.
+    rejected = [
+        c for c in captured_audit_failures
+        if c["action"] == "module.ceremony.rejected"
+    ]
+    assert len(rejected) == 1
+    payload = rejected[0]["payload"]
+    assert payload["ceremony_id"] == ceremony_id
+    assert payload["requested_action"] == "grant"
+    assert payload["reason"] == "invalid_transition"
+
+
+@pytest.mark.asyncio
+async def test_unknown_action_emits_audit_row(
+    client, captured_audit_failures
+) -> None:
+    """{"action":"banana"} → 422 + module.ceremony.rejected row via the
+    global RequestValidationError handler."""
+    clear_ceremonies()
+    email, _ = await _login_admin(client)
+
+    install = await client.post(
+        "/api/modules/install",
+        json={
+            "source": "manifest",
+            "manifest": _verified_manifest(id="legalise.r2-banana"),
+        },
+    )
+    assert install.status_code == 201
+    ceremony_id = install.json()["ceremony_id"]
+
+    resp = await client.post(
+        f"/api/modules/install/{ceremony_id}/advance",
+        json={"action": "banana"},
+    )
+    assert resp.status_code == 422
+
+    rejected = [
+        c for c in captured_audit_failures
+        if c["action"] == "module.ceremony.rejected"
+    ]
+    assert len(rejected) == 1
+    payload = rejected[0]["payload"]
+    assert payload["reason"] == "schema_validation_failed"
+    assert payload["ceremony_id"] == ceremony_id
+
+
+# ---------------------------------------------------------------------------
+# Dead-state pins (Phase 5 Step 0 carryover)
+# ---------------------------------------------------------------------------
+
+
+def test_dependency_missing_removed_from_enum() -> None:
+    """CeremonyState no longer carries DEPENDENCY_MISSING."""
+    members = {m.name for m in CeremonyState}
+    assert "DEPENDENCY_MISSING" not in members, (
+        "DEPENDENCY_MISSING was removed in Phase 5 Step 0 — see "
+        "PHASE_5_BUILD_PLAN_V3.md. Phase 4 returns 422 BEFORE start_ceremony, "
+        "so the state machine has no path to this terminal."
+    )
+    values = {m.value for m in CeremonyState}
+    assert "dependency_missing" not in values
+
+
+def test_dependency_missing_not_in_terminal_failures() -> None:
+    """The terminal-failures frozenset does not carry the dead state."""
+    from app.core.trust_ceremony import _TERMINAL_FAILURES
+
+    assert all(
+        m.value != "dependency_missing" for m in _TERMINAL_FAILURES
+    )
